데이터 타입 통합, ManyToOne 어노테이션 추가
Long -> Integer로 변경
문의 테이블에 있는 유저와 상품을 1 대 다 관계로 연결

+++ fmmall/src/main/java/com/sesac/fmmall/DTO/CartResponseDTO.java
@@ -12,7 +12,7 @@ import java.util.List;
 @AllArgsConstructor
 public class CartResponseDTO {
 
-    private Long cartId;
+    private int cartId;
     private List<CartItemResponseDTO> itemList;
     private int totalItemCount;
     private int totalPrice;
